pypi-v2/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crystalyse"
version = "1.0.1"
description = "Crystalyse v1.0 - Intelligent Scientific AI Agent for Inorganic Materials Design"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Ryan Nduma", email = "ryannduma@gmail.com"},
]
maintainers = [
    {name = "Ryan Nduma", email = "ryannduma@gmail.com"},
]
keywords = [
    "materials-design",
    "crystal-structure-prediction",
    "formation-energy",
    "computational-materials-science",
    "ai-agents",
    "chemistry",
    "inorganic-materials",
    "autonomous-agents",
    "openai-agents",
    "mcp-servers",
    "provenance",
    "computational-honesty"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.11"
dependencies = [
    # Core AI and Agent Framework
    "openai>=1.0.0",
    "openai-agents>=0.0.16",

    # Chemistry Core Libraries
    "pymatgen>=2024.1.0",
    "ase>=3.22.0",

    # AI-Powered Chemistry Tools
    "smact>=3.0.0",
    "torch>=2.1.0",
    "mace-torch>=0.3.0",
    "chemeleon-dng>=0.1.2",

    # Model Context Protocol
    "fastmcp>=0.5.0",
    "python-mcp>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.23.0",
    "starlette>=0.27.0",

    # Data Processing
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",

    # Type Safety and Validation
    "typing-extensions>=4.5.0",
    "pydantic>=2.0.0",

    # CLI and UI
    "click>=8.1.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",

    # Async and Networking
    "asyncio>=3.4.3",
    "httpx>=0.24.0",
    "websockets>=11.0.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",

    # File and Data Handling
    "PyYAML>=6.0.0",
    "toml>=0.10.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# Visualization capabilities
visualization = [
    "matplotlib>=3.7.0",
    "plotly>=5.18.0",
    "py3Dmol>=2.0.4",
    "kaleido>=0.2.1",
    "pymatviz>=0.8.5",
    "seaborn>=0.12.0",
]

# Development tools
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.11.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "build>=1.0.0",
    "twine>=4.0.0",
]

# All optional dependencies
all = [
    "crystalyse[visualization,dev]",
]

[project.urls]
Homepage = "https://github.com/ryannduma/CrystaLyse.AI"
Documentation = "https://github.com/ryannduma/CrystaLyse.AI/tree/master/docs"
Repository = "https://github.com/ryannduma/CrystaLyse.AI.git"
Issues = "https://github.com/ryannduma/CrystaLyse.AI/issues"
Changelog = "https://github.com/ryannduma/CrystaLyse.AI/blob/master/CHANGELOG.md"

[project.scripts]
crystalyse = "crystalyse.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
license-files = []

[tool.setuptools.packages.find]
where = ["src"]
include = ["crystalyse*"]
exclude = ["tests*", "docs*", "*.egg-info"]

[tool.setuptools.package-data]
"crystalyse" = [
    "prompts/*.md",
    "prompts/*.txt",
]

# Code quality tools configuration
[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["S101", "ARG", "FBT"]

[tool.ruff.isort]
known-first-party = ["crystalyse"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
